SMS Help version 1.0(1)

+++ android/project-8/src/com/example/project_8/FullInfo.java
@@ -19,7 +19,7 @@ public class FullInfo extends BasicInfo implements Serializable {
 	public String campaignType;
 	public String txtCampaign;
 	public String campaignLink;
-	public int smsSendDate;
+	public long smsSendDate;
 
 	public FullInfo(int phoneNumber, int campaignId, double priceSMS,
 			String txtSMS, String campaignName, String campaignSubName,
